Add an optional "path_prefix" field for File records
This records the "path_prefix" as used by the HERA librarian. This makes it
possible for us to re-add files to the Librarian with the correct directory
prefix within a store, while being resilient to the fact that different stores
will have different directory prefixes. Behavior should remain unchanged if
you don't specify the path_prefix.