bin/load_observations_librarian.py: get this working again
This now works again in Librarian V2.0, with much more robustness than the
previous incarnation. Most of the tricky logic has been moved to the
Librarian, which knows more about the data, so the script is now much simpler.
Files to process are fetched from the Librarian and the resulting outputs are
put back into the Librarian with the same path_prefix.