lib/dbi.py: have nodes self-identify when they're under duress
This is based on simple checks of load average, free memory, and free disk
space. This could get cleverer, of course. One idea is that nodes are better
equipped to mark themselves as being under duress -- unless of course they're
really in trouble, in which case the lack of periodic checkin will indicate to
the client that there are problems.