bin/load_observations_librarian.py: use a smarter "obsnum"
I ran into a corner case where genuinely different observations ended up with
identical "obsnums" assigned because of roundoff. But if you look in the database,
the "obsnum" is just a free text string. So just have it be the file basename,
which based on the Librarian policy will be unique for each unique file.