bin/load_observations_librarian.py: make the source configurable and re-default it
I temporarily hacked the desired data source to be "human" for testing, but this
is not what we want in production. On the other hand, this experience shows that this
value should be a command-line argument and not just hardcoded in the script.

Also fix an old exception class name. Apparently Python doesn't try to process
"except" clause definitions unless an exception is actually triggered!